requirements.txt
fastmcp
python-dotenv
pydantic
python-docx
PyPDF2
olefile
pypandoc